requirements.txt
# ── Core web framework ────────────────────────────────────────────────────────
Flask>=3.0.0
flask-cors>=4.0.0

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests>=2.31.0

# ── Environment / config ──────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── AI clients ────────────────────────────────────────────────────────────────
openai>=1.40.0

# ── Speech-to-text ────────────────────────────────────────────────────────────
faster-whisper>=1.0.0

# ── Long-term memory ─────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Gamer Mode: window capture ───────────────────────────────────────────────
pygetwindow>=0.0.9
mss>=9.0.1
Pillow>=10.0.0

# ── Smart Home (Alexa) ────────────────────────────────────────────────────────
alexapy>=1.27.0

# ── Camera vision (optional) ──────────────────────────────────────────────────
# opencv-python>=4.9.0   # Uncomment if you use camera mode
